+++ include/Utils.hpp
@@ -37,7 +37,7 @@ inline double getRandDouble01MPI()
     return dblDistr();
 }
 
-inline MPIData getMPIData(const size_t& totalNumData)
+inline MPIData getMPIData(const int_fast32_t& totalNumData)
 {
     int rank, numProcs;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
@@ -59,14 +59,14 @@ inline MPIData getMPIData(const size_t& totalNumData)
     return MPIData{ rank, numProcs, lengths, displacements };
 }
 
-inline size_t getTotalNumDataMPI(const Matrix* const data)
+inline int_fast32_t getTotalNumDataMPI(const Matrix* const data)
 {
     int rank, numProcs;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
     MPI_Comm_size(MPI_COMM_WORLD, &numProcs);
 
-    size_t totalNumData = 0;
-    size_t localNumData = data->getNumData();
+    int_fast32_t totalNumData = 0;
+    int_fast32_t localNumData = data->getNumData();
 
     MPI_Allreduce(&localNumData, &totalNumData, 1, MPI_UNSIGNED_LONG, MPI_SUM, MPI_COMM_WORLD);
